+++ common/src/main/java/net/mehvahdjukaar/dummmmmmy/client/TargetDummyModel.java
@@ -3,8 +3,10 @@ package net.mehvahdjukaar.dummmmmmy.client;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
+import com.mojang.math.Axis;
 import net.mehvahdjukaar.dummmmmmy.common.TargetDummyEntity;
 import net.mehvahdjukaar.dummmmmmy.configs.ClientConfigs;
+import net.mehvahdjukaar.moonlight.api.client.util.RotHlpr;
 import net.minecraft.client.model.HumanoidModel;
 import net.minecraft.client.model.geom.ModelPart;
 import net.minecraft.client.model.geom.PartPose;
@@ -78,7 +80,6 @@ public class TargetDummyModel<T extends TargetDummyEntity> extends HumanoidModel
                                float blue, float alpha) {
         int overlay = OverlayTexture.NO_OVERLAY;
         matrixStackIn.pushPose();
-
         this.standPlate.render(matrixStackIn, bufferIn, packedLightIn, overlay, red, green, blue, alpha);
 
         this.head.render(matrixStackIn, bufferIn, packedLightIn, overlay, red, green, blue, alpha);
@@ -107,17 +108,16 @@ public class TargetDummyModel<T extends TargetDummyEntity> extends HumanoidModel
             this.r2 = 0;
         }
 
-        // un-rotate the stand plate, so it's aligned to the block grid
-        this.standPlate.xRot = 0.0F;
-        this.standPlate.yRot = Mth.DEG_TO_RAD * -Mth.rotLerp(partialTick, entity.yRotO, entity.getYRot());
-        this.standPlate.zRot = 0.0F;
+       // un-rotate the stand plate, so it's aligned to the block grid
+       this.standPlate.xRot = 0.0F;
+       this.standPlate.yRot =  Mth.DEG_TO_RAD * -Mth.rotLerp(partialTick, entity.yBodyRotO, entity.yBodyRot);
+       this.standPlate.zRot = 0.0F;
 
     }
 
     @Override
     public void setupAnim(TargetDummyEntity entityIn, float limbSwing, float limbSwingAmount, float ageInTicks, float netHeadYaw,
                           float headPitch) {
-
 
         float n = 1.5f;